moved user testing to separate module
