+++ .github/workflows/integration_test.yaml
@@ -13,6 +13,6 @@ jobs:
       trivy-image-config: "trivy.yaml"
       juju-channel: 3.1/stable
       channel: 1.28-strict/stable
-      modules: '["test_charm", "test_saml"]'
+      modules: '["test_charm", "test_saml", "test_users"]'
       self-hosted-runner: true
       self-hosted-runner-label: "edge"
