+++ tests/integration/test_charm.py
@@ -13,9 +13,7 @@ import pytest
 import requests
 from boto3 import client
 from botocore.config import Config
-from juju.action import Action
 from juju.application import Application
-from juju.unit import Unit
 from ops.model import ActiveStatus
 from pytest_operator.plugin import Model, OpsTest
 
@@ -360,112 +358,3 @@ async def test_upgrade(
 
     await model.block_until(upgrade_finished(), timeout=10 * 60)
     check_alive()
-
-
-@pytest.mark.asyncio
-async def test_create_user(
-    app: Application,
-):
-    """
-    arrange: A discourse application
-    act: Create a user
-    assert: User is created, and re-creating the same user should fail
-    """
-
-    email = "admin-user@test.internal"
-    discourse_unit: Unit = app.units[0]
-    action: Action = await discourse_unit.run_action("create-user", email=email)
-    await action.wait()
-    assert action.results["user"] == email
-
-    # Re-creating the same user should fail, as the user already exists
-    break_action: Action = await discourse_unit.run_action("create-user", email=email)
-    await break_action.wait()
-    assert break_action.status == "failed"
-
-
-@pytest.mark.asyncio
-async def test_promote_user(
-    app: Application,
-    discourse_address: str,
-    model: Model,
-    requests_timeout: float,
-):
-    """
-    arrange: A discourse application
-    act: Promote a user to admin
-    assert: User cannot access the admin API before being promoted
-    """
-
-    def test_discourse_srv_status_ok():
-        response = requests.get(f"{discourse_address}/srv/status", timeout=requests_timeout)
-        assert response.status_code == 200
-
-    # The charm should be active when starting this test
-    await model.wait_for_idle(status="active")
-    test_discourse_srv_status_ok()
-
-    with requests.session() as session:
-        response = session.get(
-            f"{discourse_address}/session/csrf", headers={"Accept": "application/json"}, timeout=60
-        )
-        # pylint doesn't see the "ok" member
-        assert response.ok, response.text  # pylint: disable=no-member
-        data = response.json()
-        assert data["csrf"], data
-        csrf = data["csrf"]
-
-        def get_api_key() -> bool:
-            response = session.post(
-                f"{discourse_address}/admin/api/keys",
-                headers={
-                    "Content-Type": "application/json",
-                    "X-CSRF-Token": csrf,
-                    "X-Requested-With": "XMLHttpRequest",
-                },
-                json={"key": {"description": "admin-api-key", "username": None}},
-                timeout=60,
-            )
-            if response.json().get("key") is None:
-                return False
-            return True
-
-        def attempt_login(email: str, password: str):
-            response = session.post(
-                f"{discourse_address}/session",
-                headers={
-                    "Content-Type": "application/x-www-form-urlencoded; charset=UTF-8",
-                    "X-CSRF-Token": csrf,
-                    "X-Requested-With": "XMLHttpRequest",
-                },
-                data={
-                    "login": email,
-                    "password": password,
-                    "second_factor_method": "1",
-                    "timezone": "Asia/Hong_Kong",
-                },
-                timeout=60,
-            )
-            try:
-                "error" not in response.json()
-            except Exception as e:
-                logger.error("Error in response: %s", response.text)
-                raise e
-
-            assert response.ok, response.text  # pylint: disable=no-member
-            assert "error" not in response.json()
-
-        email = "test-promote-user@test.internal"
-        discourse_unit: Unit = app.units[0]
-        create_action: Action = await discourse_unit.run_action("create-user", email=email)
-        await create_action.wait()
-        assert create_action.results["user"] == email
-
-        attempt_login(email, create_action.results["password"])
-
-        assert not get_api_key(), "This should fail as the user is not promoted"
-
-        promote_action: Action = await discourse_unit.run_action("promote-user", email=email)
-        await promote_action.wait()
-
-        assert get_api_key(), "This should succeed as the user is promoted"
